% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_wetransfer.R
\name{send_wetransfer}
\alias{send_wetransfer}
\title{Creating a file transfer using the wetransfer open API}
\usage{
send_wetransfer(files_dir, api_key, name, description)
}
\arguments{
\item{files_dir}{Link to the directory where your files are stored}

\item{api_key}{The api key that you can get from https://developers.wetransfer.com}

\item{name}{The name of the transfer}

\item{description}{The description of the transfer (optional)}
}
\value{
The shortened url of the transfer
}
\description{
Creating a file transfer using the wetransfer open API
}
\examples{
send_wetransfer("path/to/files","somekey","my_first_transfer","some nice files")
}
